node identification
This commit includes a significant bit of refactoring to more
rigourously find the id for a publication node in the graph. Google go through
the effort of identifying duplicate publications on the web, and assigning
them a cluster id. So if a publication is available in a preprint repository
and on a publisher website then the single work has a unique identifier.

etudier will try to use this cluster identifier for the publication node
in the graph. Most of the time this is possible, but if a publication
has not yet been cited, and only has one version, then a cluster id is not
available in the HTML. In these cases the data-cid value on the
publication element is used. This should not have bad side effects
since by definition these are leaf nodes that are not cited.

A --debug flag was also added to get some diagnostic information as
the graph is being assembled.

fixes #1