Include year and incremental output
This commmit adds the year (when available) to the title that gets
displayed in the HTML view.

It was easier to just have étudier generate the output as it goes
instead of waiting for a captcha. The program is really iobound by
talking to Google Scholar, so I don't think there's much of a
performance impact.

closes #24
closes #7